src/app/config/settings.py
from pydantic_settings import BaseSettings
from dotenv import find_dotenv, load_dotenv

load_dotenv(find_dotenv(".env"))


class Settings(BaseSettings):
    db_uri: str = ""


app_settings = Settings()